+++ pack/__init__.py
@@ -1,16 +1,5 @@
 # -*- coding:utf-8 -*-
-formats = (
-	'SARC',
-	'DARC',
-	'BFLYT',
-	'ALYT',
-	'BFLIM',
-	'BL',
-)
-
-need_version = ('GARC',)
-need_swizzle = ('BFLIM', 'BCLIM')
-
+from ._formats import *
 
 def pack(filenames, outname, format, endian, opts):
 	print('Packing %s' % format)