+++ pack/_formats.py
@@ -1,0 +1,10 @@
+# -*- coding:utf-8 -*-
+
+formats = (
+	'SARC',
+	'DARC',
+	'BFLYT',
+	'ALYT',
+	'BFLIM',
+	'BL',
+)