+++ unpack/SARC.py
@@ -1,5 +1,5 @@
 # -*- coding:utf-8 -*-
-import os.path
+import os
 from util import error, ENDIANS
 from util.fileops import *
 from util.funcops import byterepr